fix(refund,gas): release claim on validation failure, BigDecimal sum, config validator
Review fixes from the post-refund / post-gas-funder critique:

#1 IssueRefundUseCase released the idempotency claim ONLY when the invoice
   lookup missed; if the lookup found an invoice but validation rejected
   it (status != PAID, or sum-of-refunds exceeds expected), the claim was
   left PENDING forever. The client's next attempt — with the corrected
   amount or against the right invoice — would hit IDEMPOTENCY_IN_PROGRESS
   (409) on the same key, a confusing error for a client-fixable bad
   request. Wrapped lookup + validation + custody send in a single
   pre-side-effect try block whose catch releases the claim. Only failures
   AFTER sendTransfer returns leave the claim PENDING (the intended
   manual-reconciliation case).

   Two existing test cases renamed and tightened to assert the release:
   - rejects_when_invoice_not_paid_AND_releases_claim
   - rejects_when_total_refunded_would_exceed_invoice_AND_releases_claim

#2 RefundJpaRepository.sumNonFailedAmounts declared BigInteger as the
   return type. Hibernate widens sum(NUMERIC(38,0)) to BigDecimal on
   Oracle and PostgreSQL; the BigInteger cast would have thrown
   ClassCastException at first non-zero refund total. Switched the
   query method to BigDecimal and let the repository adapter convert
   via toBigInteger(). H2 happened to return BigInteger so tests passed
   — silent runtime risk that production-shape DBs would have hit.

#6 GasFunderProperties had no validation: a misconfigured deployment
   (topUpAmount <= lowWaterMark, blank fromAddress, native asset on
   different network, duplicate funder entries, monitored address on
   an unfunded network) would fail at runtime with a baffling "balance
   below threshold but no top-up fired" symptom. GasFunderConfigValidator
   runs at bean construction time and throws IllegalStateException with
   a clear message for any of those misconfigurations — boot-time fail
   instead of runtime confusion.

DRY: PayoutMapper and RefundMapper carried identical requireNonBlank /
requireNonNull helpers built around their entity type. Extracted into
EntityFieldGuards — type-blind helpers that take a debug label so the
diagnostic still names the entity + field. Saves ~30 lines and
guarantees the two mappers report errors in the same format.

85 tests still green. ArchUnit unchanged.

+++ payment/infra/src/main/java/com/cpayment/payment/infra/persistence/jpa/JpaRefundRepository.java
@@ -9,6 +9,7 @@ import com.cpayment.payment.domain.port.RefundRepository;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
@@ -41,7 +42,7 @@ public class JpaRefundRepository implements RefundRepository {
     @Override
     @Transactional(readOnly = true)
     public BigInteger sumIssuedNonFailed(InvoiceId invoiceId) {
-        BigInteger v = jpa.sumNonFailedAmounts(invoiceId.value(), RefundStatus.FAILED);
-        return v != null ? v : BigInteger.ZERO;
+        BigDecimal v = jpa.sumNonFailedAmounts(invoiceId.value(), RefundStatus.FAILED);
+        return v != null ? v.toBigInteger() : BigInteger.ZERO;
     }
 }
