+++ payment/infra/src/main/java/com/cpayment/payment/infra/persistence/jpa/RefundJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -16,13 +16,20 @@ public interface RefundJpaRepository extends JpaRepository<RefundEntity, UUID> {
 
     List<RefundEntity> findByInvoiceId(UUID invoiceId);
 
-    /** Sum of amounts for refunds against an invoice that are NOT in {@code FAILED}. */
+    /**
+     * Sum of amounts for refunds against an invoice that are NOT in {@code FAILED}.
+     *
+     * <p>Returns {@link BigDecimal} (not {@code BigInteger}) because Hibernate widens
+     * {@code sum(NUMERIC(38,0))} to {@code BigDecimal} on Oracle / PostgreSQL — a
+     * {@code BigInteger} return type would risk {@link ClassCastException} at runtime.
+     * The repository adapter converts via {@code toBigInteger()}.
+     */
     @Query("""
         select coalesce(sum(r.amount), 0)
           from RefundEntity r
          where r.invoiceId = :invoiceId
            and r.status <> :failedStatus
         """)
-    BigInteger sumNonFailedAmounts(@Param("invoiceId") UUID invoiceId,
+    BigDecimal sumNonFailedAmounts(@Param("invoiceId") UUID invoiceId,
                                    @Param("failedStatus") RefundStatus failedStatus);
 }
